Fixed warnings in clk_axi_clkgen.c
util.h:78:7: warning: comparison of unsigned expression < 0 is always false
[-Wtype-limits]
  ((x) < 0 ? (-x) : (x))
       ^
clk_axi_clkgen.c:371:8: note: in expansion of macro 'abs'
    if (abs(f - fout) < abs(best_f - fout)) {
        ^~~
util.h:78:7: warning: comparison of unsigned expression < 0 is always false
[-Wtype-limits]
  ((x) < 0 ? (-x) : (x))
       ^
clk_axi_clkgen.c:371:24: note: in expansion of macro 'abs'
    if (abs(f - fout) < abs(best_f - fout)) {
                        ^~~
clk_axi_clkgen.c:376:16: warning: comparison between signed and unsigned
integer expressions [-Wsign-compare]
     if (best_f == fout)
                ^~

+++ drivers/axi_core/clk_axi_clkgen/clk_axi_clkgen.c
@@ -368,12 +368,12 @@ void axi_clkgen_calc_params(struct axi_clkgen *axi_clkgen,
 			dout = DIV_ROUND_CLOSEST(fvco, fout);
 			dout = clamp(dout, 1, 128);
 			f = fvco / dout;
-			if (abs(f - fout) < abs(best_f - fout)) {
+			if (abs(f - (int32_t)fout) < abs(best_f - (int32_t)fout)) {
 				best_f = f;
 				*best_d = d;
 				*best_m = m;
 				*best_dout = dout;
-				if (best_f == fout)
+				if (best_f == (int32_t)fout)
 					return;
 			}
 		}
